requirements.txt
Flask==2.3.2
flask-socketio==5.3.4
eventlet==0.33.3
gunicorn==20.1.0
pyyaml==6.0
faiss-cpu==1.7.4
openai==0.27.8
langchain==0.0.226
pandas==2.0.3
tiktoken==0.4.0



